+++ src/components/i18n/es.ts
@@ -1,0 +1,38 @@
+export const phrases = {
+  "webmapgl": {
+    "app": {
+      "menulabel": "Leyenda"
+    },
+    "basemap-switcher": {
+      "header": "Seleccione un mapa base"
+    },
+    "basemaps": {
+      "title": "Cambie mapa base"
+    },
+    "draw-toolbar": {
+      "cancel": "Cancelar",
+      "confirm": "Continuar",
+      "label": "Agrega una característica"
+    },
+    "drawer-toggle": {
+      "label": "Cambia de opcion"
+    },
+    "feature-add": {
+      "success": "Guardado exitosamente.",
+      "failure": "Se produjo un error al guardar.",
+      "label": "Agrega una característica"
+    },
+    "field": {
+      "required": "%{field} es requerido."
+    },
+    "form-page": {
+      "back": "Atrás",
+      "submit": "Salvar",
+      "cancel": "Cancelar"
+    },
+    "fullscreen": {
+      "enter": "Ingresa al modo de pantalla completa",
+      "exit": "Sal del modo de pantalla completa"
+    }
+  }
+};