+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Made gl-form-page and gl-draw-toolbar responsive.
 - Wrap text in form elements.
 - Trigger an event on gl-like-button click.
+- Wrap basemap names in switcher.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.