+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added facet class to form page.
 - Block click actions on disabled buttons.
 - Use polyglot for translation.
+- Added Spanish translation.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.