+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Block click actions on disabled buttons.
 - Use polyglot for translation.
 - Added Spanish translation.
+- Made forms translatable.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.