+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# [0.3.2] - Unreleased
+# [0.4.0] - Unreleased
 - Added zoom methods to gl-map.
 - Made gl-form-page and gl-draw-toolbar responsive.
 - Wrap text in form elements.
@@ -13,6 +13,7 @@
 - Added Spanish translation.
 - Made forms translatable.
 - Fixed map resize on drawer close.
+- Fixed an issue that caused the like button not to update.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.